feat: Pins the package manager as it's used for the first time
